Extract RollingFileReader and RollingFileWriter
Extract RollingFileReader and RollingFileWriter to the utility package.
