+++ util/build.gradle
@@ -36,6 +36,7 @@ dependencies {
   implementation 'io.vertx:vertx-core'
   implementation 'org.apache.tuweni:bytes'
   implementation 'org.apache.tuweni:units'
+  implementation 'org.xerial.snappy:snappy-java'
 
   runtimeOnly 'org.apache.logging.log4j:log4j-core'
